Move `isOriginAlive` to music/common/origin.ts
